+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 ## Change log
 
+### 8.1.2-b0 (05/08/2023 - )
+
+* search: fixed ElasticSearchLogInterceptor logging issue with chunked http entity
+  > failed to generate trace log with bulkIndex
+* search: fixed default max conn settings
+* search: updated default timeout to 15s
+  > tolerant more when es is busy
+
 ### 8.1.1 (03/14/2023 - 04/03/2023)
 
 * ws: response corresponding close-code on connect according to exception